+++ application/src/environments/environment.firebug.ts
@@ -1,6 +1,6 @@
 export const environment = {
   production: false,
-  firebug: true,
+	firebug: true,
 	gaTrackingCode: 'UA-109145893-2',
 	googleAnalytics: false
 };