+++ application/src/environments/environment.ts
@@ -1,6 +1,6 @@
 export const environment = {
   production: false,
-  firebug: false,
+	firebug: false,
 	gaTrackingCode: 'UA-109145893-2',
 	googleAnalytics: false
 };