+++ schematics/src/application/files/src/environments/environment.ts
@@ -1,4 +1,3 @@
 export const environment = {
-  production: false,
-  firebug: false
+  production: false
 };